requirements.txt
tqdm
lief
numpy
sklearn
filemagic
capstone
datasets
tokenizers
transformers
